fix(sdk): clear the sonarcloud findings this PR introduced
The segment size addition ran in int arithmetic before widening to long
(java:S2184), which is the one reliability finding failing the quality
gate. Also adds the missing @deprecated tag on DataSizeNotSupported
(java:S1123), drops an import left unused by the test changes
(java:S1128), and uses the AssertJ size assertions (java:S5838).

+++ sdk/src/main/java/io/opentdf/platform/sdk/TDF.java
@@ -498,7 +498,7 @@ class TDF {
         TDFObject tdfObject = new TDFObject();
         tdfObject.prepareManifest(tdfConfig, splits);
 
-        long encryptedSegmentSize = tdfConfig.defaultSegmentSize + kGcmIvSize + AesGcm.GCM_TAG_LENGTH;
+        long encryptedSegmentSize = (long) tdfConfig.defaultSegmentSize + kGcmIvSize + AesGcm.GCM_TAG_LENGTH;
         TDFWriter tdfWriter = new TDFWriter(outputStream);
 
         ByteArrayOutputStream aggregateHash = new ByteArrayOutputStream();
